+++ Zeze/Services/GlobalCacheManager.cs
@@ -191,7 +191,7 @@ namespace Zeze.Services
         /// </summary>
         /// <param name="p"></param>
         /// <returns></returns>
-        private int ProcessCleanup(Zeze.Net.Protocol p)
+        private long ProcessCleanup(Zeze.Net.Protocol p)
         {
             var rpc = p as Cleanup;
 
@@ -235,7 +235,7 @@ namespace Zeze.Services
             return 0;
         }
 
-        private int ProcessLogin(Zeze.Net.Protocol p)
+        private long ProcessLogin(Zeze.Net.Protocol p)
         {
             var rpc = p as Login;
             var session = Sessions.GetOrAdd(rpc.Argument.ServerId, (_) => new CacheHolder(Config));
@@ -254,7 +254,7 @@ namespace Zeze.Services
             return 0;
         }
 
-        private int ProcessReLogin(Zeze.Net.Protocol p)
+        private long ProcessReLogin(Zeze.Net.Protocol p)
         {
             var rpc = p as ReLogin;
             var session = Sessions.GetOrAdd(rpc.Argument.ServerId, (_) => new CacheHolder(Config));
@@ -267,7 +267,7 @@ namespace Zeze.Services
             return 0;
         }
         
-        private int ProcessNormalClose(Zeze.Net.Protocol p)
+        private long ProcessNormalClose(Zeze.Net.Protocol p)
         {
             var rpc = p as NormalClose;
             var session = rpc.Sender.UserState as CacheHolder;
@@ -291,7 +291,7 @@ namespace Zeze.Services
             return 0;
         }
 
-        private int ProcessAcquireRequest(Zeze.Net.Protocol p)
+        private long ProcessAcquireRequest(Zeze.Net.Protocol p)
         {
             Acquire rpc = (Acquire)p;
             if (rpc.Sender.UserState == null)
