+++ ZezeJava/src/Zeze/Net/ProtocolHandle.java
@@ -2,5 +2,5 @@ package Zeze.Net;
 
 @FunctionalInterface
 public interface ProtocolHandle {
-	int handle(Protocol p);
+	long handle(Protocol p);
 }
